+++ src/components/Sing/SequencerRuler/Container.vue
@@ -9,19 +9,12 @@
     :uiLocked
     :playheadTicks
     :sequencerSnapType
-    :isLoopEnabled
-    :loopStartTick
-    :loopEndTick
     @update:playheadTicks="updatePlayheadTicks"
     @removeTempo="removeTempo"
     @removeTimeSignature="removeTimeSignature"
     @setTempo="setTempo"
     @setTimeSignature="setTimeSignature"
     @deselectAllNotes="deselectAllNotes"
-    @setLoopEnabled="setLoopEnabled"
-    @setLoopRange="setLoopRange"
-    @clearLoopRange="clearLoopRange"
-    @addOneMeasureLoop="addOneMeasureLoop"
   />
 </template>
 
@@ -30,18 +23,12 @@ import { computed } from "vue";
 import Presentation from "./Presentation.vue";
 import { useStore } from "@/store";
 import { Tempo, TimeSignature } from "@/store/type";
-import {
-  snapTicksToGrid,
-  getTimeSignaturePositions,
-  getMeasureDuration,
-} from "@/sing/domain";
-import { baseXToTick } from "@/sing/viewHelper";
 
 defineOptions({
   name: "SequencerRuler",
 });
 
-const props = withDefaults(
+withDefaults(
   defineProps<{
     offset: number;
     numMeasures: number;
@@ -62,10 +49,6 @@ const uiLocked = computed(() => store.getters.UI_LOCKED);
 const sequencerSnapType = computed(() => store.state.sequencerSnapType);
 
 const playheadTicks = computed(() => store.getters.PLAYHEAD_POSITION);
-
-const isLoopEnabled = computed(() => store.state.isLoopEnabled);
-const loopStartTick = computed(() => store.state.loopStartTick);
-const loopEndTick = computed(() => store.state.loopEndTick);
 
 const updatePlayheadTicks = (ticks: number) => {
   void store.actions.SET_PLAYHEAD_POSITION({ position: ticks });
@@ -93,61 +76,6 @@ const removeTempo = (position: number) => {
 const removeTimeSignature = (measureNumber: number) => {
   void store.actions.COMMAND_REMOVE_TIME_SIGNATURE({
     measureNumber,
-  });
-};
-
-const setLoopEnabled = (enabled: boolean) => {
-  void store.actions.COMMAND_SET_LOOP_ENABLED({ isLoopEnabled: enabled });
-};
-
-const setLoopRange = (start: number, end: number) => {
-  void store.actions.COMMAND_SET_LOOP_RANGE({
-    loopStartTick: start,
-    loopEndTick: end,
-  });
-};
-
-const clearLoopRange = () => {
-  void store.actions.COMMAND_CLEAR_LOOP_RANGE();
-};
-
-const addOneMeasureLoop = (clickX: number) => {
-  const baseX = (props.offset + clickX) / sequencerZoomX.value;
-  const cursorTick = baseXToTick(baseX, tpqn.value);
-  const tsPositions = getTimeSignaturePositions(
-    timeSignatures.value,
-    tpqn.value,
-  );
-  const currentTs = timeSignatures.value.findLast((_, index) => {
-    return tsPositions[index] <= cursorTick;
-  });
-
-  if (!currentTs) {
-    throw new Error("assert: At least one time signature exists.");
-  }
-
-  // 1小節分のtick数を計算
-  const oneMeasureTicks = getMeasureDuration(
-    currentTs.beats,
-    currentTs.beatType,
-    tpqn.value,
-  );
-
-  // 拍子情報を考慮してスナップ
-  const startTick = snapTicksToGrid(
-    cursorTick,
-    timeSignatures.value,
-    tpqn.value,
-  );
-  const endTick = snapTicksToGrid(
-    startTick + oneMeasureTicks,
-    timeSignatures.value,
-    tpqn.value,
-  );
-
-  void store.actions.COMMAND_SET_LOOP_RANGE({
-    loopStartTick: startTick,
-    loopEndTick: endTick,
   });
 };
 </script>
