+++ src/components/Sing/SequencerRuler/LoopLane/index.ts
@@ -1,0 +1,3 @@
+import Container from "./Container.vue";
+
+export default Container;
