+++ src/styles/v2/variables.scss
@@ -34,6 +34,9 @@ $z-index-sing-background: 0;
 $z-index-sing-note: 10;
 $z-index-sing-note-lyric: 20;
 $z-index-sing-pitch: 30;
-$z-index-sing-playhead: 40;
-$z-index-sing-tool-palette: 50;
-$z-index-sing-lyric-input: 60;
+$z-index-sing-loop-background: 35;
+$z-index-sing-loop-range: 40;
+$z-index-sing-loop-handle: 45;
+$z-index-sing-playhead: 50;
+$z-index-sing-tool-palette: 60;
+$z-index-sing-lyric-input: 70;
